# src/specs/api.yaml
openapi: 3.0.0
info:
  title: API
  version: 1.0.0
  description: The main Node.js pipeline API.
  license:
    name: MIT
    url: 'https://github.com/biomage-ltd/api/blob/master/LICENSE'
  contact:
    name: Biomage Ltd.
servers:
  - url: /v1
...
paths:
...
  '/experiments/{experimentId}/samples':
    parameters:
      - schema:
          type: string
        name: experimentId
        in: path
        required: true
    get:
      summary: Your GET endpoint
      operationId: getSamplesByExperimentId
      x-eov-operation-id: samples#getSamplesByExperimentId
      x-eov-operation-handler: routes/samples
      tags: []
      responses:
        '200':
          description: 'Fetch successful, samples returned below.'
          content:
            application/json:
              schema:
                $ref: ./models/api-body-schemas/Samples.v1.yaml
        '404':
          description: Samples not found.
          content:
            application/json:
              schema:
                $ref: ./models/HTTPError.v1.yaml
      description: Get all samples for an experiment
  '/projects/{projectUuid}':
    parameters:
      - schema:
          type: string
        name: projectUuid
        in: path
        required: true
    put:
      summary: Create or update project
      operationId: updateProject
      x-eov-operation-id: projects#update
      x-eov-operation-handler: routes/projects
      responses:
        '200':
          description: Create or update sample object
          content:
            application/json:
              schema:
                $ref: ./models/HTTPSuccess.v1.yaml
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: ./models/HTTPError.v1.yaml
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: ./models/HTTPError.v1.yaml
        '415':
          description: Unsupported Media Type
          content:
            application/json:
              schema:
                $ref: ./models/HTTPError.v1.yaml
      description: Create or update a project
      requestBody:
        content:
          application/json:
            schema:
              $ref: ./models/api-body-schemas/Project.v1.yaml
        description: The new configuration to update the old one by.
    delete:
      summary: 'Deletes the project'
      operationId: deleteProject
      x-eov-operation-id: projects#delete
      x-eov-operation-handler: routes/projects
      responses:
        '200':
          description: Deletes object
          content:
            application/json:
              schema:
                $ref: ./models/HTTPSuccess.v1.yaml
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: ./models/HTTPError.v1.yaml
      description: Deletes project and associated samples.
...